requirements.txt
Package                  Version
------------------------ -----------
accelerate               1.0.1
aiohappyeyeballs         2.4.4
aiohttp                  3.10.11
aiosignal                1.3.1
async-timeout            5.0.1
attrs                    25.3.0
av                       12.3.0
bitsandbytes             0.45.5
certifi                  2025.1.31
charset-normalizer       3.4.1
datasets                 3.1.0
dill                     0.3.8
docopt                   0.6.2
filelock                 3.16.1
frozenlist               1.5.0
fsspec                   2024.9.0
huggingface-hub          0.30.1
idna                     3.10
Jinja2                   3.1.6
MarkupSafe               2.1.5
mpmath                   1.3.0
multidict                6.1.0
multiprocess             0.70.16
networkx                 3.1
num2words                0.5.14
numpy                    1.24.4
nvidia-cublas-cu12       12.1.3.1
nvidia-cuda-cupti-cu12   12.1.105
nvidia-cuda-nvrtc-cu12   12.1.105
nvidia-cuda-runtime-cu12 12.1.105
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.0.2.54
nvidia-curand-cu12       10.3.2.106
nvidia-cusolver-cu12     11.4.5.107
nvidia-cusparse-cu12     12.1.0.106
nvidia-nccl-cu12         2.20.5
nvidia-nvjitlink-cu12    12.8.93
nvidia-nvtx-cu12         12.1.105
packaging                24.2
pandas                   2.0.3
peft                     0.13.2
pillow                   10.4.0
pip                      24.2
propcache                0.2.0
protobuf                 5.29.4
psutil                   7.0.0
pyarrow                  17.0.0
python-dateutil          2.9.0.post0
pytz                     2025.2
PyYAML                   6.0.2
regex                    2024.11.6
requests                 2.32.3
safetensors              0.5.3
sentencepiece            0.2.0
setuptools               75.1.0
six                      1.17.0
sympy                    1.13.3
tokenizers               0.20.3
torch                    2.4.1
tqdm                     4.67.1
transformers             4.46.3
triton                   3.0.0
typing_extensions        4.13.1
tzdata                   2025.2
urllib3                  2.2.3
wheel                    0.44.0
xxhash                   3.5.0
yarl                     1.15.2
